requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
Pillow>=10.0.0
fal-client>=0.4.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.119.0
